+++ mento/__init__.py
@@ -8,5 +8,5 @@ Quantity = ureg.Quantity
 from .node import Node # noqa: F401, E402
 from .forces import Forces # noqa: F401, E402
 from .material import Concrete_ACI_318_19, SteelBar # noqa: F401, E402
-from .concrete.beam import RectangularConcreteBeam # noqa: F401, E402
+from .beam import RectangularBeam # noqa: F401, E402
 from .results import Formatter, TablePrinter, DocumentBuilder # noqa: F401, E402
