+++ tests/test_beam.py
@@ -1,4 +1,4 @@
-from mento.concrete.beam import RectangularConcreteBeam
+from mento.beam import RectangularBeam
 from mento.material import Concrete_ACI_318_19, SteelBar, Concrete_EHE_08
 from mento.units import psi, kip, inch, ksi, ft, mm, kN, kNm, cm, MPa
 from mento.forces import Forces
@@ -6,12 +6,12 @@ import pytest
 import numpy as np
 
 @pytest.fixture()
-def beam_example_imperial() -> RectangularConcreteBeam:
+def beam_example_imperial() -> RectangularBeam:
     concrete = Concrete_ACI_318_19(name="C4", f_c=4000*psi)  
     steelBar = SteelBar(name="ADN 420", f_y=60*ksi)  
     custom_settings = {'clear_cover': 1.5*inch, 'stirrup_diameter_ini':0.5*inch,
                        'longitudinal_diameter_ini': 1*inch} 
-    section = RectangularConcreteBeam(
+    section = RectangularBeam(
         label="V-10x16",
         concrete=concrete,
         steel_bar=steelBar,
@@ -22,17 +22,17 @@ def beam_example_imperial() -> RectangularConcreteBeam:
     return section
 
 @pytest.fixture()
-def beam_example_ehe() -> RectangularConcreteBeam:
+def beam_example_ehe() -> RectangularBeam:
     concrete= Concrete_EHE_08(name="C25",f_ck=25*MPa) 
     steelBar= SteelBar(name="B500S", f_y=500*MPa)
     custom_settings = {'clear_cover': 2.6*cm, 'stirrup_diameter_ini':8*mm,
                        'longitudinal_diameter_ini': 16*mm}
-    section = RectangularConcreteBeam(label="V-20x60",
+    section = RectangularBeam(label="V-20x60",
                                       concrete=concrete,steel_bar=steelBar,width=20*cm, height=60*cm,
                                        settings=custom_settings)
     return section
 
-def test_shear_check_EHE_08(beam_example_ehe: RectangularConcreteBeam) -> None:
+def test_shear_check_EHE_08(beam_example_ehe: RectangularBeam) -> None:
     f = Forces(V_z=100*kN, M_y=100*kNm)
     A_s = 8.04*cm**2
     beam_example_ehe.set_transverse_rebar(n_stirrups=1, d_b=6*mm, s_l=20*cm) 
@@ -54,7 +54,7 @@ def test_shear_check_EHE_08(beam_example_ehe: RectangularConcreteBeam) -> None:
     assert results.iloc[0]['Vrd,1<Vu1'] is np.True_
     assert results.iloc[0]['Vrd,2<Vu2'] is np.True_
 
-def test_shear_check_EHE_08_no_rebar_1(beam_example_ehe: RectangularConcreteBeam) -> None:
+def test_shear_check_EHE_08_no_rebar_1(beam_example_ehe: RectangularBeam) -> None:
     f = Forces(V_z=50*kN)
     A_s = 8.04*cm**2
     results = beam_example_ehe.check_shear_EHE_08(f, A_s)  
@@ -75,7 +75,7 @@ def test_shear_check_EHE_08_no_rebar_1(beam_example_ehe: RectangularConcreteBeam
     assert results.iloc[0]['Vrd,1<Vu1'] is np.True_
     assert results.iloc[0]['Vrd,2<Vu2'] is np.True_
 
-def test_shear_check_EHE_08_no_rebar_2(beam_example_ehe: RectangularConcreteBeam) -> None:
+def test_shear_check_EHE_08_no_rebar_2(beam_example_ehe: RectangularBeam) -> None:
     f = Forces(V_z=50*kN, M_y=50*kNm)
     A_s = 8.04*cm**2
     results = beam_example_ehe.check_shear_EHE_08(f, A_s)  
@@ -96,7 +96,7 @@ def test_shear_check_EHE_08_no_rebar_2(beam_example_ehe: RectangularConcreteBeam
     assert results.iloc[0]['Vrd,1<Vu1'] is np.True_
     assert results.iloc[0]['Vrd,2<Vu2'] is np.True_
 
-def test_shear_check_ACI_318_19(beam_example_imperial: RectangularConcreteBeam) -> None:
+def test_shear_check_ACI_318_19(beam_example_imperial: RectangularBeam) -> None:
     f = Forces(V_z=37.727*kip, N_x = 0*kip)  
     A_s = 0.847*inch**2 
     beam_example_imperial.set_transverse_rebar(n_stirrups=1, d_b=0.5*inch, s_l=6*inch) 
@@ -116,7 +116,7 @@ def test_shear_check_ACI_318_19(beam_example_imperial: RectangularConcreteBeam) 
     assert results.iloc[0]['Vu<ØVmax'] is np.True_
     assert results.iloc[0]['Vu<ØVn'] is np.True_
 
-def test_shear_design_ACI_318_19(beam_example_imperial: RectangularConcreteBeam) -> None:
+def test_shear_design_ACI_318_19(beam_example_imperial: RectangularBeam) -> None:
     f = Forces(V_z=37.727*kip, N_x = 0*kip) 
     A_s = 0.847*inch**2  
     results = beam_example_imperial.design_shear_ACI_318_19(f, A_s)  
@@ -140,7 +140,7 @@ def test_shear_design_ACI_318_19(beam_example_imperial: RectangularConcreteBeam)
 # Example 6.6 CRSI Design Guide
 concreteFlexureTest1 = Concrete_ACI_318_19(name="fc4000",f_c=4000*psi)
 steelBarFlexureTest1 = SteelBar(name="G60", f_y=60000*psi) 
-sectionFlexureTest1 = RectangularConcreteBeam(
+sectionFlexureTest1 = RectangularBeam(
         label="B-12x24",
         concrete=concreteFlexureTest1,
         steel_bar=steelBarFlexureTest1,
